requirements.txt
aiohttp==3.7.4
APScheduler==3.7.0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==3.0.4
cycler==0.10.0
discord.py==1.5.1
humanfriendly==9.1
idna==3.2
kiwisolver==1.3.1
matplotlib==3.3.1
multidict==4.7.6
numpy==1.20.3
Pillow==8.2.0
psutil==5.8.0
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
tzlocal==2.1
yarl==1.5.1
